fix: fixed big refactor, updated tests for missing params, etc.
