+++ clients/modelregistry_plugin/tests/test_store.py
@@ -1,2100 +1,100 @@
 """
-Tests for ModelRegistryStore
+Minimal tests for ModelRegistryTrackingStore
+
+These tests focus only on store-specific functionality:
+- Initialization and configuration
+- Basic delegation to operation classes
+- Store-specific error handling
+
+Business logic is tested in test_operations_*.py files.
+Integration is tested in test_e2e_*.py files.
 """
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 import pytest
-import requests
-from mlflow.entities import (
-    DatasetInput,
-    Experiment,
-    ExperimentTag,
-    LoggedModel,
-    LoggedModelInput,
-    LoggedModelOutput,
-    LoggedModelParameter,
-    LoggedModelStatus,
-    LoggedModelTag,
-    Metric,
-    Param,
-    Run,
-    RunInfo,
-    RunStatus,
-    RunTag,
-    ViewType,
-)
-from mlflow.exceptions import MlflowException
-from mlflow.models import Model
-from mlflow.store.entities.paged_list import PagedList
 
-from modelregistry_plugin.store import ModelRegistryStore
+from modelregistry_plugin.store import ModelRegistryTrackingStore
 
 
-class TestModelRegistryStore:
-    @pytest.fixture
-    def store(self):
-        """Create a ModelRegistryStore instance for testing."""
-        return ModelRegistryStore(
-            "modelregistry://localhost:8080", "s3://bucket/artifacts"
+class TestModelRegistryTrackingStore:
+    def test_init_default(self):
+        """Test store initialization with defaults."""
+        # Clear environment to test true defaults
+        with patch.dict("os.environ", {"MLFLOW_TRACKING_URI": ""}, clear=False):
+            store = ModelRegistryTrackingStore()
+            assert store.host == "localhost"
+            assert store.port == 8080
+            assert store.secure is False
+            assert store.base_url == "http://localhost:8080/api/model_registry/v1alpha3"
+            assert store.artifact_uri is not None
+
+    def test_init_with_custom_uri(self):
+        """Test store initialization with custom URI."""
+        store = ModelRegistryTrackingStore(
+            "modelregistry://example.com:9090", "s3://bucket/artifacts"
         )
-
-    @pytest.fixture
-    def mock_response(self):
-        """Create a mock response object."""
-        response = Mock(spec=requests.Response)
-        response.ok = True
-        response.json.return_value = {}
-        return response
-
-    @pytest.fixture
-    def mock_model(self):
-        """Create a mock MLflow Model object."""
-        model = Mock(spec=Model)
-        model.to_dict.return_value = {
-            "artifact_path": "model",
-            "run_id": "run-123",
-            "flavors": {"python_function": {}},
-            "model_uuid": "uuid-123",
-            "utc_time_created": "2023-01-01T00:00:00Z",
-            "mlflow_version": "2.0.0",
-        }
-        model_info = Mock()
-        model_info.model_uri = "runs:/run-123/model"
-        model_info.artifact_path = "model"
-        model_info.model_uuid = "uuid-123"
-        model_info.utc_time_created = "2023-01-01T00:00:00Z"
-        model_info.mlflow_version = "2.0.0"
-        model_info.flavors = {"python_function": {}}
-        model.get_model_info.return_value = model_info
-
-        # Add missing attributes that record_logged_model expects
-        model.model_id = None
-        model.model_uuid = "uuid-123"
-        model.name = "test-model"
-        model.source_run_id = "run-123"
-
-        return model
-
-    # Initialization tests
-    def test_init(self, store):
-        """Test store initialization."""
-        assert store.host == "localhost"
-        assert store.port == 8080
+        assert store.host == "example.com"
+        assert store.port == 9090
         assert store.secure is False
-        assert store.base_url == "http://localhost:8080/api/model_registry/v1alpha3"
+        assert store.artifact_uri == "s3://bucket/artifacts"
 
     def test_init_secure(self):
         """Test store initialization with secure connection."""
         with patch.dict("os.environ", {"MODEL_REGISTRY_SECURE": "true"}):
-            store = ModelRegistryStore("modelregistry://localhost:8080")
+            store = ModelRegistryTrackingStore("modelregistry://localhost:8080")
             assert store.secure is True
-
-    def test_init_with_artifact_uri(self):
-        """Test store initialization with artifact URI."""
-        store = ModelRegistryStore(
-            "modelregistry://localhost:8080", "s3://bucket/artifacts"
-        )
-        assert store.artifact_uri == "s3://bucket/artifacts"
+            assert "https://" in store.base_url
 
     def test_init_from_env(self):
         """Test store initialization from environment variable."""
         with patch.dict(
             "os.environ", {"MLFLOW_TRACKING_URI": "modelregistry://test:9090"}
         ):
-            store = ModelRegistryStore()
+            store = ModelRegistryTrackingStore()
             assert store.host == "test"
             assert store.port == 9090
 
-    # Request method tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_request_success(self, mock_request, store, mock_response):
-        """Test successful API request."""
-        mock_request.return_value = mock_response
-
-        response_data = store._request("GET", "/test")
-
-        mock_request.assert_called_once()
-        assert response_data == mock_response.json.return_value
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_request_failure(self, mock_request, store):
-        """Test failed API request."""
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = False
-        mock_response.json.return_value = {"message": "Test error"}
-        mock_response.status_code = 404
-        mock_request.return_value = mock_response
-
-        with pytest.raises(MlflowException) as exc_info:
-            store._request("GET", "/test")
-
-        assert "Model Registry API error: Test error" in str(exc_info.value)
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_request_with_custom_properties(self, mock_request, store):
-        """Test request with custom properties conversion."""
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "customProperties": {
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                }
-            }
-        }
-        mock_request.return_value = mock_response
-
-        response_data = store._request(
-            "POST", "/test", json={"customProperties": {"key1": "value1"}}
-        )
-
-        # Check that custom properties were converted in request and response
-        call_args = mock_request.call_args
-        assert (
-            call_args[1]["json"]["customProperties"]["key1"]["string_value"] == "value1"
-        )
-        assert (
-            call_args[1]["json"]["customProperties"]["key1"]["metadataType"]
-            == "MetadataStringValue"
-        )
-        assert response_data["customProperties"]["key1"] == "value1"
-
-    # Experiment operations tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_experiment(self, mock_request, store):
-        """Test experiment creation."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"id": "exp-123"}
-        mock_request.return_value = mock_response
-
-        experiment_id = store.create_experiment("test-experiment")
-
-        assert experiment_id == "exp-123"
-        # Should make 2 calls: POST to create, then PATCH to set default artifact location
-        assert mock_request.call_count == 2
-
-        # Check first call (POST to create experiment)
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiments" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["name"] == "test-experiment"
-        assert json_data["description"] == "MLflow experiment: test-experiment"
-        assert json_data["state"] == "LIVE"
-        assert json_data["customProperties"] == {}
-        # externalId should not be in the initial payload when artifact_uri is available
-        assert "externalId" not in json_data
-
-        # Check second call (PATCH to set default artifact location)
-        call_args = mock_request.call_args_list[1]
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["externalId"] == "s3://bucket/artifacts/experiments/exp-123"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_experiment_with_artifact_location(self, mock_request, store):
-        """Test experiment creation with explicit artifact_location."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"id": "exp-123"}
-        mock_request.return_value = mock_response
-
-        experiment_id = store.create_experiment(
-            "test-experiment", artifact_location="s3://custom/location"
-        )
-
-        assert experiment_id == "exp-123"
-        # Should make only 1 call since artifact_location was provided
-        assert mock_request.call_count == 1
-
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiments" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["name"] == "test-experiment"
-        assert json_data["externalId"] == "s3://custom/location"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_experiment_without_artifact_uri(self, mock_request):
-        """Test experiment creation when store has no artifact_uri."""
-        # Create store without artifact_uri
-        store = ModelRegistryStore("modelregistry://localhost:8080")
-
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"id": "exp-123"}
-        mock_request.return_value = mock_response
-        # mock response for PATCH to set default artifact location
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {}
-        mock_request.side_effect = [mock_response, mock_response_patch]
-
-        experiment_id = store.create_experiment("test-experiment")
-
-        assert experiment_id == "exp-123"
-        # Should make 2 calls: POST to create, then PATCH to set default artifact location
-        assert mock_request.call_count == 2
-
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiments" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["name"] == "test-experiment"
-        assert json_data["description"] == "MLflow experiment: test-experiment"
-        assert json_data["state"] == "LIVE"
-        assert json_data["customProperties"] == {}
-        assert json_data.get("externalId") is None
-
-        call_args = mock_request.call_args_list[1]
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["externalId"] is not None
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_experiment_with_tags(self, mock_request, store):
-        """Test experiment creation with tags."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"id": "exp-123"}
-        mock_request.return_value = mock_response
-
-        tags = [ExperimentTag("key1", "value1"), ExperimentTag("key2", "value2")]
-
-        experiment_id = store.create_experiment("test-experiment", tags=tags)
-
-        # check that the experiment id is as expected
-        assert experiment_id == "exp-123"
-
-        # Should make 2 calls: POST to create, then PATCH to set default artifact location
-        assert mock_request.call_count == 2
-
-        # Check first call (POST to create experiment)
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiments" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["name"] == "test-experiment"
-        custom_props = json_data["customProperties"]
-        assert custom_props["key1"]["string_value"] == "value1"
-        assert custom_props["key2"]["string_value"] == "value2"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_experiment(self, mock_request, store):
-        """Test getting experiment by ID."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123",
-            "customProperties": {
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        mock_request.return_value = mock_response
-
-        experiment = store.get_experiment("exp-123")
-
-        assert isinstance(experiment, Experiment)
-        assert experiment.experiment_id == "exp-123"
-        assert experiment.name == "test-experiment"
-        assert (
-            experiment.artifact_location == "s3://bucket/artifacts/experiments/exp-123"
-        )
-        assert len(experiment.tags) == 1
-        assert experiment.tags["key1"] == "value1"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_experiment_by_name(self, mock_request, store):
-        """Test getting experiment by name."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123",
-            "customProperties": {
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        mock_request.return_value = mock_response
-
-        experiment = store.get_experiment_by_name("test-experiment")
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiment" in call_args[0][1]  # endpoint
-        params = call_args[1]["params"]
-        assert params["name"] == "test-experiment"
-
-        assert isinstance(experiment, Experiment)
-        assert experiment.experiment_id == "exp-123"
-        assert experiment.name == "test-experiment"
-        assert (
-            experiment.artifact_location == "s3://bucket/artifacts/experiments/exp-123"
-        )
-        assert len(experiment.tags) == 1
-        assert experiment.tags["key1"] == "value1"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_experiment_by_name_not_found(self, mock_request, store):
-        """Test getting experiment by name when not found."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = False
-        mock_response.json.return_value = {"message": "not found"}
-        mock_response.status_code = 404
-        mock_request.return_value = mock_response
-
-        experiment = store.get_experiment_by_name("nonexistent")
-
-        assert experiment is None
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_delete_experiment(self, mock_request, store):
-        """Test deleting an experiment."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        store.delete_experiment("exp-123")
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["state"] == "ARCHIVED"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_restore_experiment(self, mock_request, store):
-        """Test restoring an experiment."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        store.restore_experiment("exp-123")
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["state"] == "LIVE"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_rename_experiment(self, mock_request, store):
-        """Test renaming an experiment."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        store.rename_experiment("exp-123", "new-name")
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["name"] == "new-name"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_experiments(self, mock_request, store):
-        """Test searching experiments."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "id": "1",
-                    "name": "exp1",
-                    "state": "LIVE",
-                    "externalId": "s3://bucket/artifacts/experiments/1",
-                    "customProperties": {
-                        "tag1": {
-                            "string_value": "value1",
-                            "metadataType": "MetadataStringValue",
-                        }
-                    },
-                },
-                {
-                    "id": "2",
-                    "name": "exp2",
-                    "state": "ARCHIVED",
-                    "externalId": "s3://bucket/artifacts/experiments/2",
-                    "customProperties": {
-                        "tag2": {
-                            "string_value": "value2",
-                            "metadataType": "MetadataStringValue",
-                        }
-                    },
-                },
-            ],
-            "nextPageToken": "token123",
-        }
-        mock_request.return_value = mock_response
-
-        result = store.search_experiments(
-            view_type=ViewType.ALL,
-            max_results=10,
-            filter_string="name='exp1'",
-            order_by=["name"],
-            page_token="token123",
-        )
-
-        assert isinstance(result, PagedList)
-        assert len(result) == 2
-        assert result[0].experiment_id == "1"
-        assert result[0].name == "exp1"
-        assert result[0].artifact_location == "s3://bucket/artifacts/experiments/1"
-        assert len(result[0].tags) == 1
-        assert result[0].tags["tag1"] == "value1"
-        assert result[1].experiment_id == "2"
-        assert result[1].name == "exp2"
-        assert result[1].artifact_location == "s3://bucket/artifacts/experiments/2"
-        assert len(result[1].tags) == 1
-        assert result[1].tags["tag2"] == "value2"
-        assert result.token == "token123"
-
-        # Verify API call parameters
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiments" in call_args[0][1]  # endpoint
-        params = call_args[1]["params"]
-        assert params["pageSize"] == 10
-        assert params["pageToken"] == "token123"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_experiments_active_only(self, mock_request, store):
-        """Test searching experiments with ACTIVE_ONLY view type."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "id": "1",
-                    "name": "exp1",
-                    "state": "LIVE",
-                    "externalId": "s3://bucket/artifacts/experiments/1",
-                    "customProperties": {},
-                },
-                {
-                    "id": "2",
-                    "name": "exp2",
-                    "state": "ARCHIVED",
-                    "externalId": "s3://bucket/artifacts/experiments/2",
-                    "customProperties": {},
-                },
-            ]
-        }
-        mock_request.return_value = mock_response
-
-        result = store.search_experiments(view_type=ViewType.ACTIVE_ONLY)
-
-        # Should only return active experiments
-        assert len(result) == 1
-        assert result[0].experiment_id == "1"
-        assert result[0].name == "exp1"
-        assert result[0].artifact_location == "s3://bucket/artifacts/experiments/1"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_experiments_deleted_only(self, mock_request, store):
-        """Test searching experiments with DELETED_ONLY view type."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "id": "1",
-                    "name": "exp1",
-                    "state": "LIVE",
-                    "externalId": "s3://bucket/artifacts/experiments/1",
-                    "customProperties": {},
-                },
-                {
-                    "id": "2",
-                    "name": "exp2",
-                    "state": "ARCHIVED",
-                    "externalId": "s3://bucket/artifacts/experiments/2",
-                    "customProperties": {},
-                },
-            ]
-        }
-        mock_request.return_value = mock_response
-
-        result = store.search_experiments(view_type=ViewType.DELETED_ONLY)
-
-        # Should only return deleted experiments
-        assert len(result) == 1
-        assert result[0].experiment_id == "2"
-        assert result[0].name == "exp2"
-        assert result[0].artifact_location == "s3://bucket/artifacts/experiments/2"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_experiments_with_filter_string(self, mock_request, store):
-        """Test searching experiments with filter string (should be ignored for now)."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"items": []}
-        mock_request.return_value = mock_response
-
-        # This should not raise an error even though filter_string is not supported yet
-        result = store.search_experiments(filter_string="name='test'")
-
-        assert isinstance(result, PagedList)
-        assert len(result) == 0
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_experiments_with_order_by(self, mock_request, store):
-        """Test searching experiments with order_by (should be ignored for now)."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"items": []}
-        mock_request.return_value = mock_response
-
-        # This should not raise an error even though order_by is not supported yet
-        result = store.search_experiments(order_by=["name"])
-
-        assert isinstance(result, PagedList)
-        assert len(result) == 0
-
-    # Run operations tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_run(self, mock_request, store):
-        """Test run creation."""
-        # Mock the raw response from Model Registry API for creating run
-        mock_response_create = Mock(spec=requests.Response)
-        mock_response_create.ok = True
-        mock_response_create.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "createTimeSinceEpoch": "1234567890",
-        }
-
-        # Mock the raw response from Model Registry API for getting experiment
-        mock_response_experiment = Mock(spec=requests.Response)
-        mock_response_experiment.ok = True
-        mock_response_experiment.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123",
-            "customProperties": {},
-        }
-
-        # Mock the raw response from Model Registry API for updating run with artifact location
-        mock_response_update = Mock(spec=requests.Response)
-        mock_response_update.ok = True
-        mock_response_update.json.return_value = {}
-
-        mock_request.side_effect = [
-            mock_response_create,
-            mock_response_experiment,
-            mock_response_update,
-        ]
-
-        run = store.create_run("exp-123", start_time=1234567890)
-
-        assert isinstance(run, Run)
-        assert run.info.run_id == "run-123"
-        assert run.info.experiment_id == "exp-123"
-        assert run.info.status == RunStatus.RUNNING
-
-        # Should make 3 calls: POST to create run, GET to get experiment, then PATCH to set artifact location
-        assert mock_request.call_count == 3
-
-        # Check first call (POST to create run)
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["experimentId"] == "exp-123"
-        assert json_data["status"] == "RUNNING"
-
-        # Check second call (GET experiment)
-        call_args = mock_request.call_args_list[1]
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-
-        # Check third call (PATCH to set artifact location)
-        call_args = mock_request.call_args_list[2]
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiment_runs/run-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert (
-            json_data["externalId"]
-            == "s3://bucket/artifacts/experiments/exp-123/run-123"
-        )
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_run_with_user_and_tags(self, mock_request, store):
-        """Test run creation with user ID and tags."""
-        # Mock the raw response from Model Registry API for creating run
-        mock_response_create = Mock(spec=requests.Response)
-        mock_response_create.ok = True
-        mock_response_create.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "owner": "user123",
-            "createTimeSinceEpoch": "1234567890",
-            "customProperties": {
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                },
-                "key2": {
-                    "string_value": "value2",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-
-        # Mock the raw response from Model Registry API for getting experiment
-        mock_response_experiment = Mock(spec=requests.Response)
-        mock_response_experiment.ok = True
-        mock_response_experiment.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123",
-        }
-
-        # Mock the raw response from Model Registry API for updating run with artifact location
-        mock_response_update = Mock(spec=requests.Response)
-        mock_response_update.ok = True
-        mock_response_update.json.return_value = {}
-
-        mock_request.side_effect = [
-            mock_response_create,
-            mock_response_experiment,
-            mock_response_update,
-        ]
-
-        tags = [RunTag("key1", "value1"), RunTag("key2", "value2")]
-
-        run = store.create_run("exp-123", user_id="user123", tags=tags)
-
-        # Should make 3 calls: POST to create run, GET to get experiment, then PATCH to set artifact location
-        assert mock_request.call_count == 3
-
-        # Check first call (POST to create run)
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["owner"] == "user123"
-        custom_props = json_data["customProperties"]
-        assert custom_props["key1"]["string_value"] == "value1"
-        assert custom_props["key2"]["string_value"] == "value2"
-
-        # check returned run has expected tag keys and values
-        assert run.data.tags.get("key1") == "value1"
-        assert run.data.tags.get("key2") == "value2"
-        assert run.info.user_id == "user123"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_run(self, mock_request, store):
-        """Test getting run by ID."""
-        # First call: get run data
-        mock_response_run = Mock(spec=requests.Response)
-        mock_response_run.ok = True
-        mock_response_run.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "name": "test-run",
-            "state": "RUNNING",
-            "owner": "user123",
-            "startTimeSinceEpoch": "1234567890",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123/run-123",
-            "customProperties": {
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        # Second call: get metrics (empty)
-        mock_response_metrics = Mock(spec=requests.Response)
-        mock_response_metrics.ok = True
-        mock_response_metrics.json.return_value = {"items": []}
-        # Third call: get params (empty)
-        mock_response_params = Mock(spec=requests.Response)
-        mock_response_params.ok = True
-        mock_response_params.json.return_value = {"items": []}
-        # Fourth call: get dataset artifacts (empty)
-        mock_response_datasets = Mock(spec=requests.Response)
-        mock_response_datasets.ok = True
-        mock_response_datasets.json.return_value = {"items": []}
-        # Fifth call: get model artifacts (empty)
-        mock_response_models = Mock(spec=requests.Response)
-        mock_response_models.ok = True
-        mock_response_models.json.return_value = {"items": []}
-
-        mock_request.side_effect = [
-            mock_response_run,
-            mock_response_metrics,
-            mock_response_params,
-            mock_response_datasets,
-            mock_response_models,
-        ]
-
-        run = store.get_run("run-123")
-
-        assert isinstance(run, Run)
-        assert run.info.run_id == "run-123"
-        assert run.info.experiment_id == "exp-123"
-        assert run.info.run_name == "test-run"
-        assert run.info.user_id == "user123"
-        assert (
-            run.info.artifact_uri == "s3://bucket/artifacts/experiments/exp-123/run-123"
-        )
-        assert len(run.data.tags) == 1
-        assert run.data.tags["key1"] == "value1"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_update_run_info(self, mock_request, store):
-        """Test updating run info."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "status": "FINISHED",
-            "endTimeSinceEpoch": "1234567899",
-            "startTimeSinceEpoch": "1234567890",
-            "owner": "user123",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123/run-123",
-        }
-        mock_request.return_value = mock_response
-
-        run_info = store.update_run_info(
-            "run-123", RunStatus.FINISHED, end_time=1234567899
-        )
-
-        assert isinstance(run_info, RunInfo)
-        assert run_info.run_id == "run-123"
-        assert run_info.status == RunStatus.FINISHED
-        assert (
-            run_info.artifact_uri == "s3://bucket/artifacts/experiments/exp-123/run-123"
-        )
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_delete_run(self, mock_request, store):
-        """Test deleting a run."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        store.delete_run("run-123")
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiment_runs/run-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["state"] == "ARCHIVED"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_restore_run(self, mock_request, store):
-        """Test restoring a run."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        store.restore_run("run-123")
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "PATCH"  # method
-        assert "/experiment_runs/run-123" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["state"] == "LIVE"
-
-    # Metric and parameter tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_log_metric(self, mock_request, store):
-        """Test logging a metric."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "metric-123",
-            "name": "accuracy",
-            "value": 0.95,
-            "step": 1,
-            "timestamp": 1234567890,
-            "createTimeSinceEpoch": 1234567890,
-            "lastModifiedTimeSinceEpoch": 1234567890,
-        }
-        mock_request.return_value = mock_response
-
-        metric = Metric("accuracy", 0.95, 1234567890, 1)
-
-        store.log_metric("run-123", metric)
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs/run-123/artifacts" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["artifactType"] == "metric"
-        assert json_data["name"] == "accuracy"
-        assert json_data["value"] == 0.95
-        assert json_data["step"] == 1
-        assert json_data["timestamp"] == "1234567890"
-        assert json_data["customProperties"] == {}
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_metric_history(self, mock_request, store):
-        """Test getting metric history for a specific metric key."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "name": "accuracy",
-                    "value": 0.95,
-                    "timestamp": "1234567890",
-                    "step": 1,
-                    "createTimeSinceEpoch": "1234567890",
-                },
-                {
-                    "name": "accuracy",
-                    "value": 0.97,
-                    "timestamp": "1234567891",
-                    "step": 2,
-                    "createTimeSinceEpoch": "1234567891",
-                },
-            ],
-            "nextPageToken": "token202",
-        }
-        mock_request.return_value = mock_response
-
-        # Get metric history for "accuracy"
-        result = store.get_metric_history("run-123", "accuracy")
-
-        # The store returns all metrics from the API response
-        # The API should filter by name, but in this test we're mocking the response
-        assert len(result) == 2
-        assert result[0].key == "accuracy"
-        assert result[0].value == 0.95
-        assert result[0].step == 1
-        assert result[1].key == "accuracy"
-        assert result[1].value == 0.97
-        assert result[1].step == 2
-
-        # Verify metrics are sorted by timestamp and step
-        assert result[0].timestamp <= result[1].timestamp
-
-        # Verify nextPageToken is handled correctly
-        assert result.token == "token202"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_metric_history_empty(self, mock_request, store):
-        """Test getting metric history when no metrics exist."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {"items": []}
-        mock_request.return_value = mock_response
-
-        metrics = store.get_metric_history("run-123", "nonexistent")
-
-        assert len(metrics) == 0
-
-        # Verify API call
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        params = call_args[1]["params"]
-        assert params["name"] == "nonexistent"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_metric_history_uses_timestamp_fallback(self, mock_request, store):
-        """Test that metric history uses createTimeSinceEpoch when timestamp is not available."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "name": "accuracy",
-                    "value": 0.95,
-                    "step": 1,
-                    "createTimeSinceEpoch": "1234567890",
-                    # Note: no "timestamp" field
-                }
-            ]
-        }
-        mock_request.return_value = mock_response
-
-        metrics = store.get_metric_history("run-123", "accuracy")
-
-        assert len(metrics) == 1
-        assert metrics[0].timestamp == 1234567890  # Should use createTimeSinceEpoch
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_metric_history_uses_timestamp_over_create_time(
-        self, mock_request, store
-    ):
-        """Test that metric history prefers timestamp over createTimeSinceEpoch."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "name": "accuracy",
-                    "value": 0.95,
-                    "timestamp": "1234567890",
-                    "step": 1,
-                    "createTimeSinceEpoch": "9999999999",  # Different value
-                }
-            ]
-        }
-        mock_request.return_value = mock_response
-
-        metrics = store.get_metric_history("run-123", "accuracy")
-
-        assert len(metrics) == 1
-        assert (
-            metrics[0].timestamp == 1234567890
-        )  # Should use timestamp, not createTimeSinceEpoch
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_metric_history_bulk_interval_from_steps(self, mock_request, store):
-        """Test basic bulk metric history retrieval for specific steps."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "name": "accuracy",
-                    "value": 0.95,
-                    "timestamp": "1234567890",
-                    "step": 1,
-                    "createTimeSinceEpoch": "1234567890",
-                },
-                {
-                    "name": "accuracy",
-                    "value": 0.98,
-                    "timestamp": "1234567892",
-                    "step": 3,
-                    "createTimeSinceEpoch": "1234567892",
-                },
-            ],
-            "nextPageToken": "token505",
-        }
-        mock_request.return_value = mock_response
-
-        # Get metrics for specific steps
-        result = store.get_metric_history_bulk_interval_from_steps(
-            "run-123", "accuracy", steps=[1, 3], max_results=2
-        )
-
-        assert len(result) == 2
-        assert all(m.run_id == "run-123" for m in result)
-        assert all(m.key == "accuracy" for m in result)
-
-        # Should return metrics for steps 1 and 3, sorted by step then timestamp
-        assert result[0].step == 1
-        assert result[0].value == 0.95
-        assert result[1].step == 3
-        assert result[1].value == 0.98
-
-        # Verify nextPageToken is handled correctly
-        assert result.token == "token505"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_log_param(self, mock_request, store):
-        """Test logging a parameter."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "param-123",
-            "name": "learning_rate",
-            "value": "0.01",
-            "parameterType": "string",
-            "createTimeSinceEpoch": 1234567890,
-            "lastModifiedTimeSinceEpoch": 1234567890,
-        }
-        mock_request.return_value = mock_response
-
-        param = Param("learning_rate", "0.01")
-
-        store.log_param("run-123", param)
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs/run-123/artifacts" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["artifactType"] == "parameter"
-        assert json_data["name"] == "learning_rate"
-        assert json_data["value"] == "0.01"
-        assert json_data["parameterType"] == "string"
-
-    # Tag management tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_set_experiment_tag(self, mock_request, store):
-        """Test setting experiment tag."""
-        # Mock the raw response from Model Registry API for GET request
-        mock_response_get = Mock(spec=requests.Response)
-        mock_response_get.ok = True
-        mock_response_get.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        # Mock the raw response from Model Registry API for PATCH request
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {}
-
-        mock_request.side_effect = [mock_response_get, mock_response_patch]
-
-        tag = ExperimentTag("key1", "value1")
-
-        store.set_experiment_tag("exp-123", tag)
-
-        # Should call PATCH to update tags
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 2  # GET + PATCH
-        patch_call = call_args[1]
-        json_data = patch_call[1]["json"]
-        custom_props = json_data["customProperties"]
-        assert custom_props["key1"]["string_value"] == "value1"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_set_tag(self, mock_request, store):
-        """Test setting run tag."""
-        # Mock the raw response from Model Registry API for GET run request
-        mock_response_get_run = Mock(spec=requests.Response)
-        mock_response_get_run.ok = True
-        mock_response_get_run.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "name": "test-run",
-            "state": "RUNNING",
-            "owner": "user123",
-            "startTimeSinceEpoch": "1234567890",
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        # Mock the raw response from Model Registry API for PATCH request
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "name": "test-run",
-            "state": "RUNNING",
-            "owner": "user123",
-            "startTimeSinceEpoch": "1234567890",
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                },
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-
-        mock_request.side_effect = [
-            mock_response_get_run,  # GET run
-            mock_response_patch,  # PATCH
-        ]
-
-        tag = RunTag("key1", "value1")
-
-        store.set_tag("run-123", tag)
-
-        # Should call PATCH to update tags
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 2  # GET run + PATCH
-        patch_call = call_args[1]
-        json_data = patch_call[1]["json"]
-        custom_props = json_data["customProperties"]
-        assert custom_props["key1"]["string_value"] == "value1"
-
-    # Batch operations tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_log_batch(self, mock_request, store):
-        """Test batch logging."""
-        # Mock the raw response from Model Registry API for GET run request
-        mock_response_get_run = Mock(spec=requests.Response)
-        mock_response_get_run.ok = True
-        mock_response_get_run.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "name": "test-run",
-            "state": "RUNNING",
-            "owner": "user123",
-            "startTimeSinceEpoch": "1234567890",
-            "customProperties": {
-                "existing": {
-                    "string_value": "tag",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        # Mock the raw response from Model Registry API for GET metrics request
-        mock_response_metrics = Mock(spec=requests.Response)
-        mock_response_metrics.ok = True
-        mock_response_metrics.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for GET params request
-        mock_response_params = Mock(spec=requests.Response)
-        mock_response_params.ok = True
-        mock_response_params.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for PATCH request
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {}
-        # Mock the raw response from Model Registry API for metric POST request
-        mock_response_metric = Mock(spec=requests.Response)
-        mock_response_metric.ok = True
-        mock_response_metric.json.return_value = {}
-        # Mock the raw response from Model Registry API for param POST request
-        mock_response_param = Mock(spec=requests.Response)
-        mock_response_param.ok = True
-        mock_response_param.json.return_value = {}
-
-        mock_request.side_effect = [
-            mock_response_get_run,  # get_run
-            mock_response_metrics,  # GET metrics
-            mock_response_params,  # GET params
-            mock_response_patch,  # PATCH for tags
-            mock_response_metric,  # POST for metric
-            mock_response_param,  # POST for param
-        ]
-
-        metrics = [Metric("acc", 0.9, 1234567890, 1)]
-        params = [Param("lr", "0.01")]
-        tags = [RunTag("new", "tag")]
-
-        store.log_batch("run-123", metrics, params, tags)
-
-        # Check that the correct number of calls were made
-        assert (
-            mock_request.call_count == 4
-        )  # GET run + PATCH + metric POST + param POST
-
-    # Input/Output logging tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_log_inputs_datasets(self, mock_request, store):
-        """Test logging dataset inputs."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {}
-        mock_request.return_value = mock_response
-
-        dataset = Mock()
-        dataset.dataset.name = "test-dataset"
-        dataset.dataset.digest = "digest123"
-        dataset.dataset.source_type = "csv"
-        dataset.dataset.source = "s3://bucket/data.csv"
-        dataset.dataset.schema = "schema"
-        dataset.dataset.profile = "profile"
-        dataset.tags = [RunTag("tag1", "value1")]
-
-        dataset_input = DatasetInput(dataset=dataset.dataset, tags=dataset.tags)
-
-        store.log_inputs("run-123", datasets=[dataset_input])
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs/run-123/artifacts" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["artifactType"] == "dataset-artifact"
-        assert json_data["name"] == "test-dataset"
-        assert json_data["digest"] == "digest123"
-        assert json_data["sourceType"] == "csv"
-        assert json_data["source"] == "s3://bucket/data.csv"
-        assert json_data["schema"] == "schema"
-        assert json_data["profile"] == "profile"
-        assert json_data["customProperties"]["tag1"]["string_value"] == "value1"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_log_inputs_models(self, mock_request, store):
-        """Test logging model inputs."""
-        # Mock the raw response from Model Registry API for GET request
-        mock_response_get = Mock(spec=requests.Response)
-        mock_response_get.ok = True
-        mock_response_get.json.return_value = {
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                }
-            }
-        }
-        # Mock the raw response from Model Registry API for POST request
-        mock_response_post = Mock(spec=requests.Response)
-        mock_response_post.ok = True
-        mock_response_post.json.return_value = {}
-
-        mock_request.side_effect = [mock_response_get, mock_response_post]
-
-        model_input = LoggedModelInput(model_id="model-123")
-
-        store.log_inputs("run-123", models=[model_input])
-
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 2  # GET + POST
-        post_call = call_args[1]
-        json_data = post_call[1]["json"]
-        custom_props = json_data["customProperties"]
-        assert custom_props["mlflow__model_io_type"]["string_value"] == "input"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_log_outputs(self, mock_request, store):
-        """Test logging model outputs."""
-        # Mock the raw response from Model Registry API for GET request
-        mock_response_get = Mock(spec=requests.Response)
-        mock_response_get.ok = True
-        mock_response_get.json.return_value = {
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                }
-            }
-        }
-        # Mock the raw response from Model Registry API for POST request
-        mock_response_post = Mock(spec=requests.Response)
-        mock_response_post.ok = True
-        mock_response_post.json.return_value = {}
-
-        mock_request.side_effect = [mock_response_get, mock_response_post]
-
-        model_output = LoggedModelOutput(model_id="model-123", step=1)
-
-        store.log_outputs("run-123", [model_output])
-
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 2  # GET + POST
-        post_call = call_args[1]
-        json_data = post_call[1]["json"]
-        custom_props = json_data["customProperties"]
-        assert custom_props["mlflow__model_io_type"]["string_value"] == "output"
-
-    # Logged model tests
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_record_logged_model(self, mock_request, store, mock_model):
-        """Test recording a logged model."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "uri": "s3://bucket/artifacts/experiments/exp-123/run-123/test-model",
-            "createTimeSinceEpoch": "1234567890",
-            "lastUpdateTimeSinceEpoch": "1234567890",
-            "artifactType": "model-artifact",
-            "state": "LIVE",
-            "customProperties": {
-                "mlflow__model_io_type": {
-                    "string_value": "output",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__source_run_id": {
-                    "string_value": "run-123",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__model_uuid": {
-                    "string_value": "uuid-123",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-        mock_request.return_value = mock_response
-
-        store.record_logged_model("run-123", mock_model)
-
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs/run-123/artifacts" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["artifactType"] == "model-artifact"
-        assert json_data["name"] == "uuid-123"
-        assert json_data["uri"] == "runs:/run-123/model"
-        assert (
-            json_data["customProperties"]["mlflow__artifactPath"]["string_value"]
-            == "model"
-        )
-        assert (
-            json_data["customProperties"]["mlflow__model_uuid"]["string_value"]
-            == "uuid-123"
-        )
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_logged_model(self, mock_request, store):
-        """Test creating a logged model."""
-        # Mock the raw response from Model Registry API for GET experiment request
-        mock_response_experiment = Mock(spec=requests.Response)
-        mock_response_experiment.ok = True
-        mock_response_experiment.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "externalId": "s3://bucket/artifacts/experiments/exp-123",
-            "customProperties": {},
-        }
-
-        # Mock the raw response from Model Registry API for POST model artifact request
-        mock_response_model = Mock(spec=requests.Response)
-        mock_response_model.ok = True
-        mock_response_model.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "uri": "s3://bucket/artifacts/experiments/exp-123/run-123/test-model",
-            "createTimeSinceEpoch": "1234567890",
-            "lastUpdateTimeSinceEpoch": "1234567890",
-            "artifactType": "model-artifact",
-            "state": "LIVE",
-            "customProperties": {
-                "mlflow__model_type": {
-                    "string_value": "sklearn",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__source_run_id": {
-                    "string_value": "run-123",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__experiment_id": {
-                    "string_value": "exp-123",
-                    "metadataType": "MetadataStringValue",
-                },
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                },
-                "param_param1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-
-        mock_request.side_effect = [mock_response_experiment, mock_response_model]
-
-        tags = [LoggedModelTag("key1", "value1")]
-        params = [LoggedModelParameter("param1", "value1")]
-
-        logged_model = store.create_logged_model(
-            "exp-123",
-            name="test-model",
-            source_run_id="run-123",
-            tags=tags,
-            params=params,
-            model_type="sklearn",
-        )
-
-        assert isinstance(logged_model, LoggedModel)
-        assert logged_model.model_id == "model-123"
-        assert logged_model.experiment_id == "exp-123"
-        assert logged_model.name == "test-model"
-        assert logged_model.source_run_id == "run-123"
-        assert logged_model.model_type == "sklearn"
-        # check logged model parameters
-        assert logged_model.params["param1"] == "value1"
-        # check logged model tags
-        assert logged_model.tags["key1"] == "value1"
-        # check logged model state
-        assert logged_model.status == LoggedModelStatus.READY
-
-        # Verify both API calls were made
-        assert mock_request.call_count == 2
-
-        # Check first call (GET experiment)
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiments/exp-123" in call_args[0][1]  # endpoint
-
-        # Check second call (POST model artifact)
-        call_args = mock_request.call_args_list[1]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs/run-123/artifacts" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["artifactType"] == "model-artifact"
-        assert json_data["name"] == "test-model"
-        assert (
-            json_data["customProperties"]["mlflow__model_type"]["string_value"]
-            == "sklearn"
-        )
-        assert (
-            json_data["customProperties"]["mlflow__experiment_id"]["string_value"]
-            == "exp-123"
-        )
-        assert (
-            json_data["customProperties"]["mlflow__source_run_id"]["string_value"]
-            == "run-123"
-        )
-        assert json_data["customProperties"]["key1"]["string_value"] == "value1"
-        assert json_data["customProperties"]["param_param1"]["string_value"] == "value1"
-        assert (
-            json_data["uri"]
-            == "s3://bucket/artifacts/experiments/exp-123/run-123/test-model"
-        )
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_logged_models(self, mock_request, store):
-        """Test searching logged models."""
-        # Mock the raw response from Model Registry API for getting runs request
-        mock_response_runs = Mock(spec=requests.Response)
-        mock_response_runs.ok = True
-        mock_response_runs.json.return_value = {
-            "items": [{"id": "run-123", "experimentId": "exp-123", "name": "test-run"}]
-        }
-
-        # Mock the raw response from Model Registry API for getting artifacts request
-        mock_response_artifacts = Mock(spec=requests.Response)
-        mock_response_artifacts.ok = True
-        mock_response_artifacts.json.return_value = {
-            "items": [{"id": "model-123", "name": "test-model"}]
-        }
-
-        # Mock the raw response from Model Registry API for getting individual logged model request
-        mock_response_model = Mock(spec=requests.Response)
-        mock_response_model.ok = True
-        mock_response_model.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "experimentId": "exp-123",
-            "uri": "s3://bucket/model",
-            "createTimeSinceEpoch": "1234567890",
-            "lastUpdateTimeSinceEpoch": "1234567890",
-            "customProperties": {
-                "mlflow__model_type": {
-                    "string_value": "sklearn",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__source_run_id": {
-                    "string_value": "run-123",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__experiment_id": {
-                    "string_value": "exp-123",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-
-        mock_request.side_effect = [
-            mock_response_runs,  # GET /experiments/exp-123/experiment_runs
-            mock_response_artifacts,  # GET /experiment_runs/run-123/artifacts
-            mock_response_model,  # GET /artifacts/model-123
-        ]
-
-        result = store.search_logged_models(
-            ["exp-123"],
-            filter_string="model_type='sklearn'",
-            max_results=10,
-            page_token="token123",
-        )
-
-        assert isinstance(result, PagedList)
-        assert len(result) == 1
-        assert result[0].model_id == "model-123"
-        assert result[0].name == "test-model"
-        assert result[0].experiment_id == "exp-123"
-        assert result[0].model_type == "sklearn"
-        assert result[0].source_run_id == "run-123"
-
-        # Verify API calls were made correctly
-        assert mock_request.call_count == 3
-
-        # Check first call: GET /experiments/exp-123/experiment_runs
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiments/exp-123/experiment_runs" in call_args[0][1]  # endpoint
-        params = call_args[1]["params"]
-        assert params["artifactType"] == "model-artifact"
-        assert params["experimentIds"] == ["exp-123"]
-        assert params["pageSize"] == "10"
-        assert params["pageToken"] == "token123"
-
-        # Check second call: GET /experiment_runs/run-123/artifacts
-        call_args = mock_request.call_args_list[1]
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiment_runs/run-123/artifacts" in call_args[0][1]  # endpoint
-        params = call_args[1]["params"]
-        assert params["artifactType"] == "model-artifact"
-
-        # Check third call: GET /artifacts/model-123
-        call_args = mock_request.call_args_list[2]
-        assert call_args[0][0] == "GET"  # method
-        assert "/artifacts/model-123" in call_args[0][1]  # endpoint
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_finalize_logged_model(self, mock_request, store):
-        """Test finalizing a logged model."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "experimentId": "exp-123",
-            "uri": "s3://bucket/model",
-            "createTimeSinceEpoch": "1234567890",
-            "lastUpdateTimeSinceEpoch": "1234567890",
-            "state": "LIVE",
-            "customProperties": {
-                "status": {
-                    "string_value": "READY",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__model_type": {
-                    "string_value": "sklearn",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-        mock_request.return_value = mock_response
-
-        logged_model = store.finalize_logged_model("model-123", LoggedModelStatus.READY)
-
-        assert isinstance(logged_model, LoggedModel)
-        assert logged_model.model_id == "model-123"
-        assert logged_model.model_type == "sklearn"
-        assert logged_model.status == LoggedModelStatus.READY
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_set_logged_model_tags(self, mock_request, store):
-        """Test setting logged model tags."""
-        # Mock the raw response from Model Registry API for GET request
-        mock_response_get = Mock(spec=requests.Response)
-        mock_response_get.ok = True
-        mock_response_get.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        # Mock the raw response from Model Registry API for PATCH request
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {}
-
-        mock_request.side_effect = [mock_response_get, mock_response_patch]
-
-        tags = [LoggedModelTag("key1", "value1"), LoggedModelTag("key2", "value2")]
-
-        store.set_logged_model_tags("model-123", tags)
-
-        # Should call PATCH to update tags
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 2  # GET + PATCH
-        patch_call = call_args[1]
-        json_data = patch_call[1]["json"]
-        custom_props = json_data["customProperties"]
-        assert custom_props["key1"]["string_value"] == "value1"
-        assert custom_props["key2"]["string_value"] == "value2"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_delete_logged_model_tag(self, mock_request, store):
-        """Test deleting a logged model tag."""
-        # Mock the raw response from Model Registry API for GET request
-        mock_response_get = Mock(spec=requests.Response)
-        mock_response_get.ok = True
-        mock_response_get.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "customProperties": {
-                "key1": {
-                    "string_value": "value1",
-                    "metadataType": "MetadataStringValue",
-                },
-                "key2": {
-                    "string_value": "value2",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-        # Mock the raw response from Model Registry API for PATCH request
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {}
-
-        mock_request.side_effect = [mock_response_get, mock_response_patch]
-
-        store.delete_logged_model_tag("model-123", "key1")
-
-        # Should call PATCH to update tags without key1
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 2  # GET + PATCH
-        patch_call = call_args[1]
-        json_data = patch_call[1]["json"]
-        custom_props = json_data["customProperties"]
-        assert "key1" not in custom_props
-        assert "key2" in custom_props
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_logged_model(self, mock_request, store):
-        """Test getting a logged model."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "uri": "s3://bucket/model",
-            "createTimeSinceEpoch": "1234567890",
-            "lastUpdateTimeSinceEpoch": "1234567890",
-            "customProperties": {
-                "mlflow__experiment_id": {
-                    "string_value": "exp-123",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__model_type": {
-                    "string_value": "sklearn",
-                    "metadataType": "MetadataStringValue",
-                },
-                "mlflow__source_run_id": {
-                    "string_value": "run-123",
-                    "metadataType": "MetadataStringValue",
-                },
-                "param_lr": {
-                    "string_value": "0.01",
-                    "metadataType": "MetadataStringValue",
-                },
-            },
-        }
-        mock_request.return_value = mock_response
-
-        logged_model = store.get_logged_model("model-123")
-
-        assert isinstance(logged_model, LoggedModel)
-        assert logged_model.model_id == "model-123"
-        assert logged_model.name == "test-model"
-        assert logged_model.experiment_id == "exp-123"
-        assert logged_model.model_type == "sklearn"
-        assert logged_model.source_run_id == "run-123"
-        assert len(logged_model.params) == 1
-        assert logged_model.params["lr"] == "0.01"
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_delete_logged_model(self, mock_request, store):
-        """Test deleting a logged model."""
-        # Mock the raw response from Model Registry API for GET request
-        mock_response_get = Mock(spec=requests.Response)
-        mock_response_get.ok = True
-        mock_response_get.json.return_value = {
-            "id": "model-123",
-            "name": "test-model",
-            "customProperties": {
-                "existing": {
-                    "string_value": "value",
-                    "metadataType": "MetadataStringValue",
-                }
-            },
-        }
-        # Mock the raw response from Model Registry API for PATCH request
-        mock_response_patch = Mock(spec=requests.Response)
-        mock_response_patch.ok = True
-        mock_response_patch.json.return_value = {}
-
-        mock_request.side_effect = [mock_response_get, mock_response_patch]
-
-        store.delete_logged_model("model-123")
-
-        # Should call PATCH to mark as archived
-        call_args = mock_request.call_args_list
-        assert len(call_args) == 1  # PATCH only
-        patch_call = call_args[0]
-        json_data = patch_call[1]["json"]
-        assert json_data["artifactType"] == "model-artifact"
-        custom_props = json_data["customProperties"]
-        assert custom_props["state"]["string_value"] == "MARKED_FOR_DELETION"
-
-    # Search runs test
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_runs_all(self, mock_request, store):
-        """Test searching runs."""
-        # Mock the raw response from Model Registry API for search request
-        mock_response_search = Mock(spec=requests.Response)
-        mock_response_search.ok = True
-        mock_response_search.json.return_value = {
-            "items": [
-                {
-                    "id": "run-123",
-                    "experimentId": "exp-123",
-                    "name": "test-run",
-                    "state": "RUNNING",
-                    "owner": "user123",
-                    "startTimeSinceEpoch": "1234567890",
-                    "externalId": "s3://bucket/artifacts/experiments/exp-123/run-123",
-                    "customProperties": {
-                        "tag1": {
-                            "string_value": "value1",
-                            "metadataType": "MetadataStringValue",
-                        }
-                    },
-                }
-            ],
-            "nextPageToken": "token123",
-        }
-        # Mock the raw response from Model Registry API for metrics request
-        mock_response_metrics = Mock(spec=requests.Response)
-        mock_response_metrics.ok = True
-        mock_response_metrics.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for params request
-        mock_response_params = Mock(spec=requests.Response)
-        mock_response_params.ok = True
-        mock_response_params.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for dataset artifacts request
-        mock_response_datasets = Mock(spec=requests.Response)
-        mock_response_datasets.ok = True
-        mock_response_datasets.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for model artifacts request
-        mock_response_models = Mock(spec=requests.Response)
-        mock_response_models.ok = True
-        mock_response_models.json.return_value = {"items": []}
-
-        mock_request.side_effect = [
-            mock_response_search,  # search request
-            mock_response_metrics,  # metrics request
-            mock_response_params,  # params request
-            mock_response_datasets,  # dataset artifacts request
-            mock_response_models,  # model artifacts request
-        ]
-
-        result = store.search_runs(
-            ["exp-123"],
-            filter_string="status='RUNNING'",
-            max_results=10,
-            page_token="token123",
-        )
-
-        assert isinstance(result, PagedList)
-        assert len(result) == 1
-        assert result[0].info.run_id == "run-123"
-        assert (
-            result[0].info.artifact_uri
-            == "s3://bucket/artifacts/experiments/exp-123/run-123"
-        )
-        assert result.token is None
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_runs_active_only(self, mock_request, store):
-        """Test searching runs with ACTIVE_ONLY view type."""
-        # Mock the raw response from Model Registry API for search request
-        mock_response_search = Mock(spec=requests.Response)
-        mock_response_search.ok = True
-        mock_response_search.json.return_value = {
-            "items": [
-                {
-                    "id": "run-123",
-                    "experimentId": "exp-123",
-                    "name": "active-run",
-                    "status": "RUNNING",  # MLflow RunStatus
-                    "state": "LIVE",  # ModelRegistry lifecycle state
-                    "owner": "user123",
-                    "startTimeSinceEpoch": "1234567890",
-                    "externalId": "s3://bucket/artifacts/experiments/exp-123/run-123",
-                    "customProperties": {},
-                },
-                {
-                    "id": "run-456",
-                    "experimentId": "exp-123",
-                    "name": "deleted-run",
-                    "status": "FINISHED",  # MLflow RunStatus
-                    "state": "ARCHIVED",  # ModelRegistry lifecycle state
-                    "owner": "user123",
-                    "startTimeSinceEpoch": "1234567890",
-                    "externalId": "s3://bucket/artifacts/experiments/exp-123/run-456",
-                    "customProperties": {},
-                },
-            ],
-            "nextPageToken": "token123",
-        }
-        # Mock the raw response from Model Registry API for metrics request (for active run)
-        mock_response_metrics_active = Mock(spec=requests.Response)
-        mock_response_metrics_active.ok = True
-        mock_response_metrics_active.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for params request (for active run)
-        mock_response_params_active = Mock(spec=requests.Response)
-        mock_response_params_active.ok = True
-        mock_response_params_active.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for dataset artifacts request (for active run)
-        mock_response_datasets_active = Mock(spec=requests.Response)
-        mock_response_datasets_active.ok = True
-        mock_response_datasets_active.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for model artifacts request (for active run)
-        mock_response_models_active = Mock(spec=requests.Response)
-        mock_response_models_active.ok = True
-        mock_response_models_active.json.return_value = {"items": []}
-
-        mock_request.side_effect = [
-            mock_response_search,  # search request
-            mock_response_metrics_active,  # metrics request for active run
-            mock_response_params_active,  # params request for active run
-            mock_response_datasets_active,  # dataset artifacts request for active run
-            mock_response_models_active,  # model artifacts request for active run
-            # Note: No nested calls for deleted run since it gets filtered out
-        ]
-
-        result = store.search_runs(["exp-123"], run_view_type=ViewType.ACTIVE_ONLY)
-
-        # Should return all runs since filtering is not implemented yet
-        assert isinstance(result, PagedList)
-        assert len(result) == 1
-        assert result[0].info.run_id == "run-123"
-        assert (
-            result[0].info.artifact_uri
-            == "s3://bucket/artifacts/experiments/exp-123/run-123"
-        )
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_search_runs_deleted_only(self, mock_request, store):
-        """Test searching runs with DELETED_ONLY view type."""
-        # Mock the raw response from Model Registry API for search request
-        mock_response_search = Mock(spec=requests.Response)
-        mock_response_search.ok = True
-        mock_response_search.json.return_value = {
-            "items": [
-                {
-                    "id": "run-123",
-                    "experimentId": "exp-123",
-                    "name": "active-run",
-                    "status": "RUNNING",  # MLflow RunStatus
-                    "state": "LIVE",  # ModelRegistry lifecycle state
-                    "owner": "user123",
-                    "startTimeSinceEpoch": "1234567890",
-                    "externalId": "s3://bucket/artifacts/experiments/exp-123/run-123",
-                    "customProperties": {},
-                },
-                {
-                    "id": "run-456",
-                    "experimentId": "exp-123",
-                    "name": "deleted-run",
-                    "status": "FINISHED",  # MLflow RunStatus
-                    "state": "ARCHIVED",  # ModelRegistry lifecycle state
-                    "owner": "user123",
-                    "startTimeSinceEpoch": "1234567890",
-                    "externalId": "s3://bucket/artifacts/experiments/exp-123/run-456",
-                    "customProperties": {},
-                },
-            ],
-            "nextPageToken": "token123",
-        }
-        # Mock the raw response from Model Registry API for metrics request (for deleted run)
-        mock_response_metrics_deleted = Mock(spec=requests.Response)
-        mock_response_metrics_deleted.ok = True
-        mock_response_metrics_deleted.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for params request (for deleted run)
-        mock_response_params_deleted = Mock(spec=requests.Response)
-        mock_response_params_deleted.ok = True
-        mock_response_params_deleted.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for dataset artifacts request (for deleted run)
-        mock_response_datasets_deleted = Mock(spec=requests.Response)
-        mock_response_datasets_deleted.ok = True
-        mock_response_datasets_deleted.json.return_value = {"items": []}
-        # Mock the raw response from Model Registry API for model artifacts request (for deleted run)
-        mock_response_models_deleted = Mock(spec=requests.Response)
-        mock_response_models_deleted.ok = True
-        mock_response_models_deleted.json.return_value = {"items": []}
-        """Test error handling in requests."""
-        # Mock the raw response from Model Registry API with error
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = False
-        mock_response.text = "Server error"
-        mock_response.json.return_value = {"message": "Invalid JSON"}
-        mock_response.status_code = 400
-        mock_request.return_value = mock_response
-
-        with pytest.raises(MlflowException) as exc_info:
-            store._request("GET", "/test")
-
-        assert "Model Registry API error: Invalid JSON" in str(exc_info.value)
-
-    def test_get_artifact_location(self, store):
-        """Test getting artifact location."""
-        response_with_external_id = {"externalId": "s3://bucket/artifacts"}
-        response_without_external_id = {}
-
-        assert (
-            store._get_artifact_location(response_with_external_id)
-            == "s3://bucket/artifacts"
-        )
-
-        # When no externalId is provided, the store uses its artifact_uri as fallback
-        assert (
-            store._get_artifact_location(response_without_external_id)
-            == "s3://bucket/artifacts"
-        )
-
-        # Test with a store that has no artifact_uri
-        store_no_artifact = ModelRegistryStore("modelregistry://localhost:8080")
-        assert (
-            store_no_artifact._get_artifact_location(response_without_external_id)
-            is not None  # uses default file store location instead of None
-        )
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_create_run_without_artifact_uri(self, mock_request):
-        """Test run creation when experiment has no artifact_location."""
-        # Create store without artifact_uri
-        store = ModelRegistryStore("modelregistry://localhost:8080")
-
-        # Mock the raw response from Model Registry API for creating run
-        mock_response_create = Mock(spec=requests.Response)
-        mock_response_create.ok = True
-        mock_response_create.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "createTimeSinceEpoch": "1234567890",
-        }
-
-        # Mock the raw response from Model Registry API for getting experiment (no artifact_location)
-        mock_response_experiment = Mock(spec=requests.Response)
-        mock_response_experiment.ok = True
-        mock_response_experiment.json.return_value = {
-            "id": "exp-123",
-            "name": "test-experiment",
-            "customProperties": {},
-        }
-
-        # Mock the raw response from Model Registry API for patching run
-        mock_response_run_patch = Mock(spec=requests.Response)
-        mock_response_run_patch.ok = True
-        mock_response_run_patch.json.return_value = {
-            "id": "run-123",
-            "experimentId": "exp-123",
-            "createTimeSinceEpoch": "1234567890",
-            "externalId": "file:./mlruns/run-123",
-        }
-
-        mock_request.side_effect = [
-            mock_response_create,
-            mock_response_experiment,
-            mock_response_run_patch,
-        ]
-
-        run = store.create_run("exp-123", start_time=1234567890)
-
-        assert isinstance(run, Run)
-        assert run.info.run_id == "run-123"
-        assert run.info.experiment_id == "exp-123"
-        assert run.info.status == RunStatus.RUNNING
-        assert run.info.artifact_uri is not None
-
-        # Should make 3 calls using experiment default artifact_location
-        assert mock_request.call_count == 3
-
-        # Check first call (POST to create run)
-        call_args = mock_request.call_args_list[0]
-        assert call_args[0][0] == "POST"  # method
-        assert "/experiment_runs" in call_args[0][1]  # endpoint
-        json_data = call_args[1]["json"]
-        assert json_data["experimentId"] == "exp-123"
-        assert json_data["status"] == "RUNNING"
-        assert json_data.get("externalId") is None
-
-    @patch("modelregistry_plugin.store.requests.request")
-    def test_get_metric_history_basic(self, mock_request, store):
-        """Test basic metric history retrieval."""
-        # Mock the raw response from Model Registry API
-        mock_response = Mock(spec=requests.Response)
-        mock_response.ok = True
-        mock_response.json.return_value = {
-            "items": [
-                {
-                    "name": "accuracy",
-                    "value": 0.95,
-                    "timestamp": "1234567890",
-                    "step": 1,
-                    "createTimeSinceEpoch": "1234567890",
-                },
-                {
-                    "name": "accuracy",
-                    "value": 0.97,
-                    "timestamp": "1234567891",
-                    "step": 2,
-                    "createTimeSinceEpoch": "1234567891",
-                },
-            ],
-            "nextPageToken": "token202",
-        }
-        mock_request.return_value = mock_response
-
-        result = store.get_metric_history("run-123", "accuracy")
-
-        assert len(result) == 2
-        assert all(metric.key == "accuracy" for metric in result)
-        assert result[0].value == 0.95
-        assert result[0].step == 1
-        assert result[0].timestamp == 1234567890
-        assert result[1].value == 0.97
-        assert result[1].step == 2
-        assert result[1].timestamp == 1234567891
-
-        # Verify API call
-        mock_request.assert_called_once()
-        call_args = mock_request.call_args
-        assert call_args[0][0] == "GET"  # method
-        assert "/experiment_runs/run-123/metric_history" in call_args[0][1]  # endpoint
-        params = call_args[1]["params"]
-        assert params["name"] == "accuracy"
-
-        # Verify nextPageToken is handled correctly
-        assert result.token == "token202"
+    def test_operation_classes_initialized(self):
+        """Test that all operation classes are properly initialized."""
+        store = ModelRegistryTrackingStore("modelregistry://localhost:8080")
+
+        # Verify operation classes exist
+        assert hasattr(store, "experiments")
+        assert hasattr(store, "runs")
+        assert hasattr(store, "metrics")
+        assert hasattr(store, "models")
+        assert hasattr(store, "search")
+
+        # Verify they have the same API client
+        assert store.experiments.api_client is store.api_client
+        assert store.runs.api_client is store.api_client
+        assert store.metrics.api_client is store.api_client
+        assert store.models.api_client is store.api_client
+        assert store.search.api_client is store.api_client
+
+    def test_basic_delegation_smoke_test(self):
+        """Smoke test that store methods delegate to operation classes."""
+        store = ModelRegistryTrackingStore("modelregistry://localhost:8080")
+
+        # Test a few key methods exist and are callable
+        # We don't test the actual implementation - that's in operation tests
+        assert callable(store.create_experiment)
+        assert callable(store.get_experiment)
+        assert callable(store.create_run)
+        assert callable(store.get_run)
+        assert callable(store.log_metric)
+        assert callable(store.log_param)
+        assert callable(store.search_runs)
+        assert callable(store.search_experiments)
+
+    def test_api_client_configuration(self):
+        """Test that API client is properly configured."""
+        store = ModelRegistryTrackingStore("modelregistry://localhost:8080")
+
+        assert store.api_client.base_url == store.base_url
+        assert hasattr(store.api_client, "session")
+        assert hasattr(store.api_client, "request")
 
 
 if __name__ == "__main__":
